docshell/test/navigation/test_bug1583110.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>test bug 1583110</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script>
    SimpleTest.waitForExplicitFinish();
    var bc = new BroadcastChannel("bug1583110");
    var pageshowCount = 0;
    bc.onmessage = function(event) {
      ok(event.data.type == "pageshow");
      ++pageshowCount;
      if (pageshowCount == 1) {
        bc.postMessage("loadnext");
      } else if (pageshowCount == 2) {
        bc.postMessage("back");
      } else {
        ok(event.data.persisted, "Should have persisted the first page");
        bc.close();
        SimpleTest.finish();
      }
    }

    function test() {
      window.open("file_bug1583110.html", "", "noopener");
    }
  </script>
</head>
<body onload="test()">
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
</body>
</html>
